Upversion to 1.6.
Update the description in install.rdf to match the changes described
below.

Major structural change: To avoid all the issues I was trying to work
around with switching the view while the mouse is hovering over the
folder list, require the mouse to hover over the view title
instead. This is unfortunately necessary until/unless drag & drop
functionality in Gecko becomes richer.

Hide the pref for enabling / disabling the arrows, since they now need
to be enabled or there's no place for the mouse to hover. Also hide
the pref for the secondary delay, because there is no longer a
secondary delay.

Set up a recurring timer to check if the session is gone and if so
revert to the cached folder view. Necessary so that the view will
revert if the drag & drop is canceled, e.g., if the user hits ESC or
presses a mouse button.