Thunderbird 48+ compatibility: folder pane toolbar, no more arrows
As of Thunderbird 48, there's now a folder pane toolbar and the arrows
are gone, so our code for enabling the display of the arrows is gone
as well. Instead, you can enable the folder pane toolbar and then
customize it to add the "Folder Views" tool to switch between folder
views.

The remainder of the add-on's functionality, i.e., automatically
switching to the "All Folders" view when you hover over the toolbar
while dragging a message, remains the same.