Put folder pane view switcher arrows back in Thunderbird 49+
The folder pane view switcher arrows were removed in Thunderbird 49
and replaced with an invisible-by-default folder pane toolbar into
which one could optionally put a pop-up menu to select the current
view.

The Folder Pane View Switcher add-on was fixed after that so that its
automatic "Switch to 'All Folders' view on drag-and-drop"
functionality worked with the new folder pane toolbar, but the arrows
were still gone.

This change puts the arrows back, with the option to hide them as an
add-on preference (the preference is set to make them visible by
default).

This change also makes the folder pane toolpar visible automatically
if it isn't already visible.