Add support for choosing which views are active
Increment version number to 2.0.

The preferences dialog now lists all available views and allows the
user to select whether each view is accessible in the pop-up menu, and
if so, whether it is also in the rotation of views selectable using
the arrows.

This version of the add-on only supports Thunderbird 49 or newer. If
you want to use an older Thunderbird version, you will need to use
version 1.14, which does not have the view selection functionality
described above.

This commit also fixes a separate preferences-related bug that was
present for Thunderbird 59 and newer, which aren't yet generally
available: preferences weren't being displayed or updated properly in
the preferences dialog.