Fix usage of destructuring assignment to fix default preferences loading
I misunderstood how JavaScript destructuring assignment works, and as
a result, my code for loading the default preferences values was
incorrect. I'm not sure this has any functional impact because I think
it's possible that Thunderbird is loading default preferences itself,
but the code should be correct in any case, so I've fixed it.