FIX: Make it easier to remove views from arrows
It's supposed to be possible for a particular view to be available in
the drop-down menu but not included in the rotation that gets cycled
through with the errors. It _was_ possible to do that, but only if you
opened the preferences, unchecked the menu checkbox for the view, and
then checked it again, at which point you could uncheck the arrows
checkbox if you wanted to. Now the arrows checkbox is editable
immediately when the preferences open, as long as the menu checkbox is
checked.